add rename

+++ src/main/java/ru/kostapo/cloudfilestorage/service/MinIoService.java
@@ -1,7 +1,9 @@
 package ru.kostapo.cloudfilestorage.service;
 
+import io.lettuce.core.ScriptOutputType;
 import io.minio.messages.Item;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Service;
 import ru.kostapo.cloudfilestorage.entity.dto.MinIoReqObject;
 import ru.kostapo.cloudfilestorage.entity.dto.MinIoResObject;
@@ -11,6 +13,7 @@ import ru.kostapo.cloudfilestorage.repository.MinioRepository;
 import java.util.ArrayList;
 import java.util.List;
 
+@Log4j2
 @Service
 @RequiredArgsConstructor
 public class MinIoService implements StorageService {
@@ -19,12 +22,14 @@ public class MinIoService implements StorageService {
 
     @Override
     public List<MinIoResObject> getAllObjectsByFolder(String username, String folderPath) {
+        log.info("GET ALL BY FOLDER");
         List<Item> items = minioRepository.getAllByFolder(username, folderPath);
         return ObjectMapper.INSTANCE.mapItemsToMinIoResObjects(items);
     }
 
     @Override
     public List<MinIoResObject> getAllObjectsBySearchQuery(String username, String query) {
+        log.info("FIND ALL BY SEARCH QUERY");
         List<MinIoResObject> result = new ArrayList<>();
         List<Item> items = minioRepository.getAllByUser(username);
         for (Item item : items) {
@@ -38,6 +43,7 @@ public class MinIoService implements StorageService {
 
     @Override
     public void uploadFiles(List<MinIoReqObject> files, String path) {
+        log.info("UPLOAD FILES");
         for (MinIoReqObject dto : files) {
             minioRepository.uploadFile(dto.getHolder(), path, dto.getFile());
         }
@@ -45,24 +51,38 @@ public class MinIoService implements StorageService {
 
     @Override
     public void uploadFolder(String username, String path, String name) {
+        log.info("UPLOAD FOLDER");
         minioRepository.uploadFolder(username, path, name);
     }
 
     @Override
     public void deleteObject(String username, MinIoResObject object) {
         if (object.isItIsDir()) {
+            log.info("DELETE FOLDER");
             minioRepository.removeFolder(username, object.getFullPath() + object.getObjectName());
         } else {
+            log.info("DELETE FILE");
             minioRepository.removeFile(username, object.getFullPath() + object.getObjectName());
         }
     }
 
     @Override
     public void renameObject(String username, MinIoResObject object, String newName) {
-        String oldSource = object.getFullPath() + object.getObjectName();
-        String newSource = object.getFullPath() + newName;
-        minioRepository.copyObject(username, oldSource, newSource);
-        deleteObject(username, object);
+        System.out.println("NEW NAME: "  + newName);
+        if (object.isItIsDir()) {
+            log.info("RENAME FOLDER");
+            List<Item> objectsForRename = minioRepository.getAllByPath(username, object.getFullPath() + object.getObjectName());
+            for(Item i : objectsForRename) {
+                MinIoResObject tmp = ObjectMapper.INSTANCE.itemToMinIoResObject(i);
+                if(!isObjectDir(i)) {
+                    replaceFolderName(username, tmp, object.getObjectName(), newName);
+                }
+            }
+            deleteObject(username, object);
+        } else {
+            log.info("RENAME FILE");
+            replaceFileName(username, object, newName);
+        }
     }
 
     private String getObjectName(String objectPath) {
@@ -73,5 +93,24 @@ public class MinIoService implements StorageService {
             }
         }
         return "";
+    }
+
+    private boolean isObjectDir(Item item) {
+        System.out.println("check: " + item.isDir() + " + " + item.objectName());
+        return !item.isDir() && item.objectName().endsWith("/");
+    }
+
+    private void replaceFileName(String username, MinIoResObject file, String newName) {
+        String oldSource = file.getFullPath() + file.getObjectName();
+        String newSource = file.getFullPath() + newName;
+        minioRepository.copyObject(username, oldSource, newSource);
+        deleteObject(username, file);
+    }
+
+    private void replaceFolderName(String username, MinIoResObject file, String oldName, String newName) {
+        String oldPath = file.getFullPath() + file.getObjectName();
+        String newPath = oldPath.replace(oldName, newName);
+        minioRepository.copyObject(username, oldPath, newPath);
+        deleteObject(username, file);
     }
 }
